note_app_backend/notes/serializers.py
from rest_framework import serializers
from .models import Note


class NoteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Note
        fields = ['id', 'title', 'description', 'recording_file', 'created_at', 'user']
        read_only_fields = ['user']
